% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_bins}
\alias{calc_bins}
\title{Generate bins for continous variables}
\usage{
calc_bins(df, metric, bin.threshold = 4, min.bins = 3, max.bins = 8,
  bin.cutoff = 0.055)
}
\arguments{
\item{df}{A dataframe containing metric values}

\item{metric}{The column name of the metric of interest}

\item{bin.threshold}{The min number of bins that needs to be met in order to calculate bins up to max.bins}

\item{min.bins}{The min number of bins}

\item{max.bins}{The max number of bins}

\item{bin.cutoff}{Combines bins together where the percent of values contained is less than the cutoff}
}
\value{
A list object containing the input df with outliers removed, and the outlier values
}
\description{
This function calculates outliers for any given metric
}
\examples{
calc_bins(df, 'spend')
}
