% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_efficiency}
\alias{calc_efficiency}
\title{Calculate weighting efficiency for a given set of weights}
\usage{
calc_efficiency(x, base = 1)
}
\arguments{
\item{x}{A vector of weights}

\item{base}{Base weights (default = 1)}
}
\value{
The weighting efficiency score
}
\description{
This function calculates weighting efficiency, which is an indication of the amount of skewing that had to be done to get the weights to converge
}
\examples{
calc_efficiency(weights)
}
