% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_moe}
\alias{calc_moe}
\title{Calculate margin of error}
\usage{
calc_moe(x, B, n)
}
\arguments{
\item{x}{A vector of values}

\item{B}{Number of resampling iterations}

\item{n}{Resample size}
}
\value{
A list containing various metrics
}
\description{
This function calculates margin of error using bootstrap resampling
}
\examples{
moe.summary <- calc_moe()
moe.summary$interval
moe.summary$mean
moe.summary$moe
}
