% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_outlier}
\alias{calc_outlier}
\title{Calculate outliers}
\usage{
calc_outlier(x, both = FALSE)
}
\arguments{
\item{x}{A vector containing values}
}
\value{
A scalar value of the outlier
}
\description{
This function calculates outliers for any given metric
}
\examples{
x <- c(1000, seq(1,100, 1))
calc_outlier(x) # returns 1000
}
