% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_outliers}
\alias{calc_outliers}
\title{Calculate outliers}
\usage{
calc_outliers(df, metric)
}
\arguments{
\item{df}{A dataframe containing metric values}

\item{metric}{The column name of the metric of interest}
}
\value{
A list object containing the input df with outliers removed, and the outlier values
}
\description{
This function calculates outliers for any given metric
}
\examples{
calc_outliers(df, 'spend')
}
