% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_props}
\alias{calc_props}
\title{Calculate proportions}
\usage{
calc_props(df, exclude.cols)
}
\arguments{
\item{df}{A dataframe containing variables to calculate proportions from}

\item{exclude.cols}{A vector of columns to be excluded}
}
\value{
A dataframe with variabes, levels, and proportions
}
\description{
This function calculates variable proportions for each column in a dataframe
}
\examples{
calc_props(df, 'user_id')
}
