% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_weights}
\alias{calc_weights}
\title{Generate weights for a given set of users}
\usage{
calc_weights(df, y, min.cap = -Inf, max.cap = Inf)
}
\arguments{
\item{df}{A dataframe containing metric values}

\item{y}{A df that contains quotas (i.e. targets) that will be used to determine weights}

\item{min}{Lower cap weight floor. Defaults to -Inf}

\item{max}{Upper cap weight ceiling. Defaults to Inf}
}
\value{
A list object containing the input df with weights attached, and proportions for each variable used
}
\description{
This function calculates weights for a set of users, based on predefined targets
}
\examples{
targets <- calc_props(df.test, 'user_id')
df.list <- calc_weights(df.control, targets)
df.weighted <- df.list$df
}
