% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{copy_local_tbl}
\alias{copy_local_tbl}
\title{Upload local .csv to a Vertica table}
\usage{
copy_local_tbl(conn, table.name, file.path)
}
\arguments{
\item{conn}{A database connection object}

\item{table.name}{A character string of the schema.table name}

\item{file.path}{A chracter string of the local file path (see example)}
}
\value{
A message that informs you of when the copy is completed
}
\description{
This function loads CSV-formatted data into a Vertica table, using COPY statements
}
\examples{
vertica <- vertica_connect(driver.path)
table.name <- 'dev.test_table'
file.path <- '/Users/johndoe/Desktop/test_data.csv'
copy_local_tbl(vertica, table.name, file.path)
}
