% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{create_tbl}
\alias{create_tbl}
\title{Create a Vertica Table}
\usage{
create_tbl(conn, table.name, table.vars)
}
\arguments{
\item{conn}{A database connection object}

\item{table.name}{A character string of the schema.table name}

\item{table.vars}{A chracter string with SQL table column definitions (see example)}
}
\value{
A message that informs you of when the table is successfully created
}
\description{
This function creates a table on Vertica, while dropping any existing table if exists
}
\examples{
vertica <- vertica_connect(driver.path)
table.name <- 'dev.test_table'
table.vars <- 'user_id INT, trips INT, gender VARCHAR'
create_tbl(vertica, table.name, table.vars)
}
