% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_bm_static}
\alias{get_bm_static}
\title{Pull a B&M static}
\usage{
get_bm_static(conn, static_end_date = "'2017-12-31'")
}
\arguments{
\item{conn}{A database connection object}

\item{static_end_date}{A SQL-formatted month-end date (see examples)}
}
\value{
A dataframe containing users and other details
}
\description{
This function returns a list of users for any given 12-month B&M static
}
\examples{
vertica <- vertica_connect(driver.path)
# single static
df <- get_bm_static(vertica, static_end_date="'2017-12-31'") 
# for multiple statics
df <- get_bm_static(vertica, static_end_date=c("'2017-12-31', '2016-12-31'") 
}
