% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{recode_demos}
\alias{recode_demos}
\title{Recode raw demographics (used for weighting)}
\usage{
recode_demos(df)
}
\arguments{
\item{df}{A dataframe that contains values to be recoded}
}
\value{
The input dataframe with appended recoded vars
}
\description{
This function recodes demographics to those labels used in demogrpahic weighting.
Requires: gender_id, age, income_id, ethnicity_id, has_children_id, hh_size_id
}
\examples{
df <- recode_demos(df)
}
