% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{snowflake_connect}
\alias{snowflake_connect}
\title{Connect to Snowflake}
\usage{
snowflake_connect(driver.path = Sys.getenv("SNOWFLAKE_DRIVER_PATH"),
  user = Sys.getenv("SNOWFLAKE_USER"), pw = Sys.getenv("SNOWFLAKE_PW"))
}
\arguments{
\item{driver.path}{Local path to the Snowflake JDBC driver (.jar)}

\item{user}{Database username}

\item{pw}{Database password}
}
\value{
A Snowflake connection object
}
\description{
This function creates a Snowflake JDBC connection object for querying InfoScout tables directly from R.
}
\examples{
driver.path = '/users/me/documents/drivers/snowflake-jdbc-3.6.6.jar'
conn <- snowflake_connect(driver.path)
df <- dbGetQuery(conn, 'SELECT * FROM us_tlog.d_user LIMIT 10;')
}
