% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{vertica_connect}
\alias{vertica_connect}
\title{Connect to Vertica}
\usage{
vertica_connect(driver.path, user, pw)
}
\arguments{
\item{driver.path}{Local path to the Vertica JDBC driver (.jar)}

\item{user}{Database username}

\item{pw}{Database password}
}
\value{
A Vertica connection object
}
\description{
This function creates a Vertica JDBC connection object for querying InfoScout tables directly from R.
}
\examples{
driver.path = '/users/me/documents/drivers/vertica-jdbc-7.2.3-0.jar'
vertica <- vertica_connect(driver.path)
df <- dbGetQuery(vertica, 'SELECT * FROM us_tlog.d_user LIMIT 10;')
}
